Optimization and UI update
- Adjusted layout
    - Removed Drawer
    - Moved sorting shelf to bottom horizontally
    - Removed recycle button
- Reworded start screen and added buttons for arrangement selection
- Arrangement selection now spawns in 1 tool at a time

- Adjusted dial to be able to move only left and right
    - Can no longer go from 1 to 10 or 10 to 1 from moving it past the limit
- Added player choice for how they want to sort minerals
- Added tutorialization for tools
- Adjusted collision code to not give errors on simulation end
    - Respawn zone
        - No more parenting
        - Simplified respawn
    - Drop zone
        - No more parenting
        - Smoother and less complicated lerping to proper position
- Added restart button to win screen to loop easily
- Added rules to mineral selection on respawning them so that each property has a unique value being used

+++ Assets/Assets/Scripts/MineralScriptableObject.cs
@@ -9,7 +9,7 @@ public class MineralScriptableObject : ScriptableObject
     
     [Header("Attributes")]
     [Tooltip("Based off Moh's Scale")]
-    public float hardness;
+    public int hardness;
     
     [Tooltip("1 = Cubic\n" +
              "2 = Tetragonal\n" +
